Revert "Make `-F winapp` buildable on all platforms (#33)"
This reverts commit ab51f023009761298e7fb92646df8a0648578032.
